requirements/common.txt
# Django>1.8,<1.9
django_autofixture
django-debug-toolbar
# PyMySQL
pytz
